requirements.txt
pycrypto==2.6.1
pycryptodomex==3.4.5
codeclimate-test-reporter==0.2.1
nose==1.3.7
pylint==1.6.5
flake8==3.3.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
git+https://github.com/romanvm/Kodistubs.git#egg=Kodistubs
httpretty==0.8.14
mock==1.0.1
requests==2.12.4
pydes==2.0.1
radon==2.1.1
Sphinx==1.5.5
sphinx_rtd_theme==0.2.4
m2r==0.1.12
git+https://github.com/asciidisco/kodi-release-helper.git#egg=kodi-release-helper
dennis==0.9
blessings==1.6
demjson==2.2.4
yamllint==1.8.1
restructuredtext_lint==1.1.1
